Moved the JavaScript file to the Plugins folder
