+++ frontend/src/plugins/router.js
@@ -1,7 +1,7 @@
-import Home from './components/Home.vue'
-import Profile from './components/users/Profile.vue'
-import Registration from './components/sessions/Registration.vue'
-import Login from './components/sessions/Login.vue'
+import Home from '../components/Home.vue'
+import Profile from '../components/users/Profile.vue'
+import Registration from '../components/sessions/Registration.vue'
+import Login from '../components/sessions/Login.vue'
 const routes = [
   {
     component: Home,
